Cast usipy_msg_heap_remaining to size_t.

+++ src/usipy_msg_heap.h
@@ -17,6 +17,6 @@ void *usipy_msg_heap_alloc(struct usipy_msg_heap *, size_t);
 )
 
 #define usipy_msg_heap_remaining(hp) \
-  ((hp)->tsize - (hp)->alen)
+  (size_t)((hp)->tsize - (hp)->alen)
 
 #endif /* _USIPY_MSG_HEAP_H */
